[FEATURE] New class cluster for groups of similar junctions
This PR adds a new class cluster that represents a cluster of junctions by holding an std::vector<junction>. The new class implements public methods get_average_mate1() and get_average_mate2 to obtain the average of all junctions in the cluster. The PR also updates the simple clustering method to use the new cluster class. The PR resolves #51.